+++ paymentmethod/HeidelpayPayPal.php
@@ -6,6 +6,7 @@ namespace Plugin\s360_unzer_shop5\paymentmethod;
 
 use JTL\Checkout\Bestellung;
 use JTL\Shop;
+use Plugin\s360_unzer_shop5\src\Payments\Traits\HasSavedPaymentData;
 use UnzerSDK\Resources\PaymentTypes\BasePaymentType;
 use UnzerSDK\Resources\TransactionTypes\AbstractTransactionType;
 use UnzerSDK\Resources\TransactionTypes\Charge;
@@ -42,6 +43,7 @@ class HeidelpayPayPal extends HeidelpayPaymentMethod implements RedirectPaymentI
     use HasAuthorization;
     use HasDirectCharge;
     use HasMetadata;
+    use HasSavedPaymentData;
 
     /**
      * Although Paypal support both auth as well as charge calls, we only support Direct Charge.
@@ -55,6 +57,9 @@ class HeidelpayPayPal extends HeidelpayPaymentMethod implements RedirectPaymentI
     {
         /** @var Config $config */
         $config = Shop::Container()->get(Config::class);
+        if ($config->getPaymentSetting(Config::ALLOW_SAVE, $this->moduleID) === 'Y') {
+            $this->savePaymentData($payment);
+        }
 
         // Create a customer with shipping address for Paypal's Buyer Protection
         $shopCustomer = $this->sessionHelper->getFrontendSession()->getCustomer();
@@ -95,7 +100,8 @@ class HeidelpayPayPal extends HeidelpayPaymentMethod implements RedirectPaymentI
             $this->createCharge($order),
             $payment->getId(),
             $customer,
-            $this->createMetadata()
+            $this->createMetadata(),
+            $basket
         );
     }
 }