+++ paymentmethod/HeidelpayPrepayment.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Plugin\s360_unzer_shop5\paymentmethod;
 
+use Plugin\s360_unzer_shop5\src\Payments\Traits\HasBasket;
 use UnzerSDK\Resources\PaymentTypes\BasePaymentType;
 use UnzerSDK\Resources\TransactionTypes\AbstractTransactionType;
 use UnzerSDK\Resources\TransactionTypes\Charge;
@@ -30,6 +31,7 @@ use Plugin\s360_unzer_shop5\src\Payments\Traits\HasMetadata;
  */
 class HeidelpayPrepayment extends HeidelpayPaymentMethod
 {
+    use HasBasket;
     use HasMetadata;
     use HasCustomer;
 
@@ -100,6 +102,16 @@ class HeidelpayPrepayment extends HeidelpayPaymentMethod
             $this->debugLog('Updated Customer Resource: ' . $customer->jsonSerialize(), static::class);
         }
 
+        // Create Basket
+        $session = $this->sessionHelper->getFrontendSession();
+        $basket = $this->createHeidelpayBasket(
+            $session->getCart(),
+            $order->Waehrung,
+            $session->getLanguage(),
+            $order->cBestellNr ?? $payment->getId()
+        );
+        $this->debugLog('Basket Resource: ' . $basket->jsonSerialize(), static::class);
+
         $charge = new Charge(
             $this->getTotalPriceCustomerCurrency($order),
             $order->Waehrung->getCode(),
@@ -111,7 +123,8 @@ class HeidelpayPrepayment extends HeidelpayPaymentMethod
             $charge,
             $payment->getId(),
             $customer,
-            $this->createMetadata()
+            $this->createMetadata(),
+            $basket
         );
     }
 }