+++ src/Controllers/FrontendOutputController.php
@@ -6,6 +6,7 @@ namespace Plugin\s360_unzer_shop5\src\Controllers;
 
 use JTL\Checkout\Bestellung;
 use JTL\Shop;
+use Plugin\s360_unzer_shop5\src\Services\SavedPaymentDataService;
 use Plugin\s360_unzer_shop5\src\Utils\Config;
 
 /**
@@ -17,12 +18,26 @@ class FrontendOutputController extends Controller
 {
     private const TEMPLATE_ID_CHANGE_PAYMENT_METHOD = 'template/partials/change_payment';
     private const TEMPLATE_ID_PAYMENT_INFO = 'template/partials/payment_info';
+    private const TEMPLATE_ID_SAVED_PAYMENT_DATA = 'template/partials/saved_payment_methods';
 
     /**
      * @inheritDoc
      */
     public function handle(): string
     {
+        // Show Saved payment data
+        if (Shop::getPageType() == \PAGE_MEINKONTO) {
+            $pqMethod = $this->config->get(Config::PQ_METHOD_SAVED_PAYMENT_DATA, 'after');
+            $pqSelector = $this->config->get(Config::PQ_SELECTOR_SAVED_PAYMENT_DATA, '.account-data-item-orders');
+            $paymentData = (new SavedPaymentDataService())->getUserPaymentData();
+
+            if (!empty($paymentData) && $pqSelector) {
+                pq($pqSelector)->{$pqMethod}(
+                    $this->view(self::TEMPLATE_ID_SAVED_PAYMENT_DATA, ['s360_unzer' => ['savedPaymentMethods' => $paymentData]])
+                );
+            }
+        }
+
         // Add "Change Payment Button"/Link
         if (Shop::getPageType() == \PAGE_BESTELLVORGANG) {
             $snippet = $this->view(self::TEMPLATE_ID_CHANGE_PAYMENT_METHOD);